Massive update.jpeg RED: Red has proven itself an oppressive force once again. This slew of nerfs should keep it in check.
- Hypervoxel, DrKig now end the turn.
Aurich sets the standard that boardwipes that summon fighters should end the turn. HyperVoxel has very little counterplay and Kig makes the game meainingless once cast should you follow it up immediately.

- Ninja Elite, Avote, Nymn have reduced stats
Nymn and Avote are extremely powerful bread cards. The hope is that shaving off stats from them should keep bread in line and using Avote on itself is more discouraged.

- DerptasticDevil Corrupt Guest find it harder to generate red
It is starting to become a rule that red cards that cost all white are inherently too strong. By adding red icons into these costs, it is hoped that this makes red fairer and harder to generate.

- Antisammeh now destroys fighters with 200 or less health instead of dealing damage
Anitsammeh is a borderline pernicious card. It is hoped that this nerf makes his bread far more awkward and difficult to use.

- Mag has reduced stats
The 25 health was to make mag harder to wipe. At 1 red, this was a huge mistake. Mag is now less awkward to deal with.

- Rinicye and MrDoomBringer have reworked cards
One of the better counters to red is to play giant walls. Rinicye and Doombringer completely nullify this interaction; Enemy Swapping is officailly no longer a red effect.

OTHER:
- FixingFailure can't generate icons, has reduced health
Once each large update, I am going to nerf one monowhite card by making it unable to generate icons until the deck is safe. Monowhite is an intriguing way to play the game but has proven itself degenerate in the past few months.

- FabulousBench's effect is nerfed, has reduced health
~ Delirious Ooze is completely redesigned
One issue Blox cards is facing is the presense of "I win cards" that render the game up until that point completely meaningless. FabulousBench and DeliriousOoze are the two most obvious examples, and as such have took a hit.

+ Zerax's effect is slightly buffed
This is an attempt to diverisfy the amount of green decks and increase the number of viable finishers to make various strains of green control different.

+ Dukran hsa a health buff
Dukran was intended to be a flashy card that easily sets up alars. This has not happened, and a health buff is an attempt at allowing this.

MINOR:
Grand Chicken Lord's bio cleaned up
Myrmiredon alt bio changed
Wishnite's bio changed
Cheshire Cat, Eisenhower effect attempted to be fixed